doit: Split out cmd_probe
